попытка 3

+++ src/main/java/com/example/finaltask/repository/AvatarRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvatarRepository extends JpaRepository<UserAvatar,Long> {
+public interface AvatarRepository extends JpaRepository<UserAvatar,Integer> {
 
 //    UserAvatar getById();
